feat: Add blocked table to database schema
This commit adds a new table called "blocked" to the database schema. The "blocked" table is used to store information about users who have been blocked by other users. It has columns for the user ID and the blocked user ID, along with a timestamp for when the block was created. This table will be used to implement the blocking functionality in the application.